requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
openai==1.3.0
sqlalchemy==2.0.23
pydantic==2.5.0
<<<<<<< HEAD
stripe==7.0.0
flask
gunicorn
=======
stripe==5.4.0

gunicorn==22.0.0
>>>>>>> 15018256b5eac04f96868a9d266c3091629a84a8
